Fix solution to part 2 of day 16
Python lets you alter the list that you are iterating over. That is very kind and very annoying.
Apparently, when an item is removed, the iteration skips the next item (because everything is moved?)
So when 2 invalid tickets occur consecutively, I would remove the first, but not the second.
Creating a copy fixes this.